Commit primeiro dia

+++ Projetos/Aula3/ex009.py
@@ -1,2 +1,2 @@
-m=int(input('Digite o valor em metros: '))
+m=float(input('Digite o valor em metros: '))
 print(f'O valor que em metros é: {m} \nO valor em centimetros é: {m*100} \nE o valor em milimetros é: {m*1000}')
